requirements.txt
Django==4.0.1
djangorestframework==3.13.1
psycopg2>=2.8.6,<2.8.7
django-filter==0.15.3

